# api.yaml
swagger: '2.0'
info:
  title: API микросервиса авторизации
  version: 1.0.0
paths:
...
  /auth/v1/private/groups:
    get:
      tags:
        - EmployeeGroup
      description: Получить список элементов справочника групп пользователей без учета прав
      operationId: getEmployeeGroupListWithOutPermissions
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/employeeGroupRs'
        403:
          description: Forbidden
        500:
          description: Internal server error
  /auth/v1/groups:
    get:
      tags:
        - EmployeeGroup
      description: Запрос получения списка записей справочника групп пользователей с пегинацией
      operationId: getEmployeeGroupList
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            description: Постраничный вывод групп пользователей
            $ref: '#/definitions/groupData'
        403:
          description: Forbidden
        500:
          description: Internal server error
    post:
      tags:
        - EmployeeGroup
      description: Создание записи в справочнике групп пользователей
      operationId: createGroup
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: groupData
          description: Данные записи справочника групп пользователя
          schema:
            $ref: '#/definitions/groupData'
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/groupData'
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/groups/{id}:
    get:
      tags:
        - EmployeeGroup
      description: Получение информации о группе пользователей
      operationId: getGroupDataById
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: Идентификатор группу пользователей
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: ОК
          schema:
            description: Информация о группе пользователей
            $ref: '#/definitions/groupData'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
    put:
      tags:
        - EmployeeGroup
      description: Обновление информации о группе пользователей
      operationId: updateGroupById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: Идентификатор группы пользователей
          required: true
          type: integer
          format: int64
        - in: body
          name: groupData
          description: Новые данные о группе пользователей
          schema:
            $ref: '#/definitions/groupData'
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
...
definitions:
...
  employeeGroupRs:
    type: object
    properties:
      groups:
        description: Список всех записей справочника групп пользователей
        type: array
        items:
          $ref: '#/definitions/groupData'
  groupData:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи справочника группы пользователя
      name:
        type: string
        description: Наименование записи справочника группы пользователя
      code:
        type: string
        description: Код группы записи справочника
  errorData:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            fieldName:
              type: string
              description: Поле, к которому относится ошибка
            errorMessage:
              type: string
              description: Сообщение об ошибке
...